requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
email-validator==2.2.0
reportlab==4.0.7
